gh-155. Fix api tests discussions
