+++ apps/web-portal/package.json
@@ -18,9 +18,8 @@
     "clean:node-modules": "rimraf node_modules"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "confme": "^1.1.0",
     "body-parser": "^1.19.0",
+    "confme": "^1.1.0",
     "cors": "^2.8.5",
     "express": "^4.17.1"
   }
