+++ packages/util-web-api-client/lib/axios/AxiosAuthApi.mjs
@@ -13,6 +13,8 @@ export class AxiosAuthApi extends AuthApi {
   }
 
   async login(params) {
-    return this.#http.post('/api/v1/auth/login', params)
+    const response = await this.#http.post('/api/v1/auth/login', params)
+
+    return response ? response.data : {status: 1}
   }
 }
